+++ proofs/compiler/x86_instr_decl.v
@@ -196,6 +196,7 @@ Variant x86_op : Type :=
 | VPCLMULQDQ of wsize
 
 (* SHA instructions *)
+| SHA256RNDS2
 | SHA256MSG1
 | SHA256MSG2
 .
@@ -2105,6 +2106,12 @@ Definition Ox86_VPCLMULQDQ_instr :=
 
 (* -------------------------------------------------------------------------------------- *)
 (* SHA instructions *)
+Definition Ox86_SHA256RNDS2_instr :=
+  mk_instr_pp "SHA256RNDS2" (w3_ty U128) (w_ty U128)
+    [:: Eu 0; Eu 1; ADExplicit (AK_mem Unaligned) 2 (ACR_vector XMM0)]
+    [:: Eu 0] MSB_MERGE sha256rnds2
+    [:: [:: xmm; xmmm true; xmm ]] 3 (primM SHA256RNDS2) (pp_name_ty "sha256rnds2" [:: U128; U128; U128 ]).
+
 Definition Ox86_SHA256MSG1_instr :=
   mk_instr_pp "SHA256MSG1" (w2_ty U128 U128) (w_ty U128) [:: Eu 0; Eu 1] [:: Eu 0] MSB_MERGE sha256msg1
     (check_xmm_xmmm U128) 2 (primM SHA256MSG1) (pp_name_ty "sha256msg1" [::U128;U128]).
@@ -2265,6 +2272,7 @@ Definition x86_instr_desc o : instr_desc_t :=
   | VAESKEYGENASSIST   => Ox86_VAESKEYGENASSIST_instr.1
   | PCLMULQDQ          => Ox86_PCLMULQDQ_instr.1
   | VPCLMULQDQ sz      => Ox86_VPCLMULQDQ_instr.1 sz
+  | SHA256RNDS2        => Ox86_SHA256RNDS2_instr.1
   | SHA256MSG1         => Ox86_SHA256MSG1_instr.1
   | SHA256MSG2         => Ox86_SHA256MSG2_instr.1
   end.
@@ -2421,6 +2429,7 @@ Definition x86_prim_string :=
    Ox86_VAESKEYGENASSIST_instr.2;
    Ox86_PCLMULQDQ_instr.2;
    Ox86_VPCLMULQDQ_instr.2;
+   Ox86_SHA256RNDS2_instr.2;
    Ox86_SHA256MSG1_instr.2;
    Ox86_SHA256MSG2_instr.2;
    ("VPMAX"%string,